feat: update ALE-Pong-v5 configuration and benchmark script
- Modified `ALE-Pong-v5.yaml` to adjust `ent_coef`, `n_steps`, and `eval_freq_epochs`, enhancing training parameters for improved performance.
- Introduced `batch_size` parameter to `ALE-Pong-v5.yaml` for better control over training dynamics.
- Refactored `benchmark_vecenv_fps.py` for improved readability by restructuring environment creation logic.
- These changes aim to optimize training configurations and enhance code clarity, facilitating better training outcomes and usability.

+++ utils/trainer_factory.py
@@ -36,5 +36,6 @@ def build_trainer(
         limit_val_batches=limit_val_batches,
         num_sanity_val_steps=0,
         log_every_n_steps=log_every_n_steps,
-        callbacks=callbacks
+        callbacks=callbacks,
+        precision="16-mixed"
     )
